requirements.txt
cryptography==46.0.3
django-allauth==65.13.0
oauthlib==3.3.1
pillow==12.0.0
pip-chill==1.0.3
psycopg2-binary==2.9.11
pyjwt==2.10.1
python-dotenv==1.2.1
requests==2.32.5
